requirements.txt
pytorch_lightning==2.4.0
gdown>=4.6.0
tyro==0.8.11
scipy==1.14.1
torch-fidelity
flask
matplotlib